feat: re-enable Exa search docs: add information regarding what models to use in `.env.example`
From my testing, Mistral was the best at telling me the current Android + iOS versions. It actually used DuckDuckGo. However, it was still slightly off for Android. Might have been better to only provide Exa instead. In the future, it may be beneficial to add an argument that disables DuckDuckGo or just disable DuckDuckGo if Exa is being used.